+++ src/components/proposals/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './useSignAndSubmitProposal';
+export * from './usePrepareAndSignVoteData';
